Removed assigning the ID to a field
Assigning an ID to a field is a special case of using the “userFields”. The need to assign an ID is highly depending on the underlying catalogue and field configuration and may vary from none (if provided from within the catalogue) to many (in some situations). Therefore, the “userFields” should be used to assign data.