Bug fix for Bug #1233136
The task of calculating the ATS or—in absence of an author—the TSL is a
trivial task which doesn’t depend on anything else than those two
strings. There is no need to define this function in an interface and
implement it over and over again in various plugins.

• Added public static method to ProzesskopieForm
• Rewrote method for better readability and to use StringBuilder
(faster)
• Corrected CalcProzesstitel() to work without catalogue
• Removed unnecessary method from interface IOpacPlugin
• Removed unnecessary method from PicaOpacImport source
• Updated picaOpacImportClass.jar in plugins directory