messages: Add English plural form of "Benutzer" and use it
German "Benutzer" is the same word for singular and plural, while the
English translation is either "User" or "Users", so we need two different
texts for it.

The menu entry must be "Users" instead of "User", and the caption of
the corresponding page must also be "Users". Fix this and translate
the comments for both code parts to English.

Signed-off-by: Stefan Weil <sw@weilnetz.de>