Fix GPL text (FSF address was wrong) - part 2
The address of the FSF changed many years ago. Instead of using the
current address, the new text was copied from GPL 3 which simply
redirects to a URL. This avoids unnecessary future changes when the
FSF moves again.

Here some files are fixed which were recently added with the wrong address.

Signed-off-by: Stefan Weil <sw@weilnetz.de>