Fix “Synchronization on Boolean”
The code synchronized on a Boolean. Since there normally exist only two
Boolean objects in the whole VM, this code could be synchronizing on the
same object as other, unrelated code, leading to unresponsiveness and
possible deadlock.

In addition, the method synchronized on a non-final field which changed
value. synchronized() gets a lock on the referenced object, not on the
variable, so the code could not provide the expected mutual exclusion.