Fix incorrect lazy initialization and update of static field
The method performed an unsynchronized lazy initialization of the static
field “config”. After the field was set, the object was further updated,
but the field is visible to other threads as soon as it is set. It might
be better to not set the static field until the value fully initialized.