Fixes Bug: EmptyTask.clone() does not call super.clone()
The function is a misnomer. Theads are like matchsticks for one-time use
only and cannot be cloned. A burnt down thread needs to be replaced,
this is what the implementation already does, so the function was
renamed to make clear its behaviour differs from the contract of
java.lang.Object#clone().