Fix GPL text (FSF address was wrong) - part 3
The address of the FSF changed many years ago. Instead of using the
current address, the new text was copied from GPL 3 which simply
redirects to a URL. This avoids unnecessary future changes when the
FSF moves again.

Fix all remaining files which were not handled by the previous patches.

Signed-off-by: Stefan Weil <sw@weilnetz.de>