Change log level to INFO and truncate overcomplicated message
If MetadataTypeNotAllowedException is thrown in this place it this is a
case that is intended to happen, because an optional field was tried to
add. If this field is not allowed here, it is just the correct behaviour
of the program not to add it. Since some logging is intended in cases
where an exception is thrown the exception is retargeted to INFO level
to state that the program is functioning correctly.

The error message that was logged is overly detailed for our case:

> Couldn’t add (MetadataType) to (DocStrctType): No corresponding
> MetadataType object in result of DocStruc.getAllMetadataTypes().

Therefore a replacement was put in place to make it more concise:

> Couldn’t add optional field (MetadataType) to (DocStrctType).