Fix breaking exports if data copy rules cannot be applied
The design idea of data copy rules is that rules that cannot be applied
for a certain process are skipped silently. This provides the ability to
add various rules which may be in use in different contexts. If a rule
cannot be applied an exception will be thrown during application of the
rule. These exceptions are caught on a per-rule basis.