Pass only the ID of the newspaper batch to the worker thread
Asynchronous work should be passed the ID of the object only. Found
here: http://blog.schauderhaft.de/2008/09/28/hibernate-sessions-in-two-tier-rich-client-applications/

> The answer is actually fairly easy once found: […] Background tasks
get their own session. The critical point is the transfer of objects
from one frame to another. It is tempting to just pass the object but
then you have a object from the wrong session in the frame. Instead pass
just the Id (the primary key) and use that to load the object in the new
session. This approach solved most of our problems nicely.