Fix Javadoc warnings
This patch fixes 81 Javadoc warnings.

While changing the Javadoc comments, the patch also tries to improve some
of them a little bit or removes documentation which obviously no longer
applies. Fix also typos found in the same context.

In one case, a German variable name was changed to the English name
which was already used in Javadoc.

Signed-off-by: Stefan Weil <sw@weilnetz.de>