Add .gitignore paths required to seemlessly integrate in Eclipse
Integrating kitodo-production as an Eclipse project will create the
additional files `.classpath` and `.project`, as well as a `.settings/`
directory.

The `Goobi` directory is used as the web content directory. To satisfy
the `GoobiVersionListener`, a `META-INF/` directory must be added in the
web content directory.

To deploy the configuration for the local instance, a copy of the config
files, modified to match the local configuration, must be placed in
`Goobi/src/`.