Access the UGH though the interface
The access to the UGH classes is redirected through the previously
developed interface. In this way, in the next step, the UGH can be
replaced by a replacement implementation of the interface.

To access the UGH, use UghImplementation.INSTANCE from
org.kitodo.legacy. Do not mention the legacy UGH
classes in the source code any more.