Delete processes from batches that are templates
If someone ever linked a process template into a batch, which was possible in the front-end, this is a mistake in the data that needs to be fixed in order for database migration to work. Batches are were not inteded to ever contain templates and cannot contain templates from here on. Those batches cannot ever have been useful for anything, so this can be safely deleted.