Inserts action links to create images and implements the functionality
Functioning:
To manually start image generation, corresponding action links are
displayed in the task. If you click on one of the links, the generation
will be triggered. There are tree different functions that trigger the
creation of all images, the creation of missing images, or the creation
of damaged and missing images.

Implementation:
The additional action links have been added to the taskBoxActivities.xml
page. The functions have been implemented in CurrentTaskForm. Each of
them parameterizes an ImageGenerator, which is passed to a
TaskImageGeneratorThread for execution in the task manager, since the
creation of the images can take a while for large processes. This allows
an operator to manage the progress of the task in TaskManager. The class
ImageGeneratorTask has been implemented. For this a function had to be
developed, which can list the contents of a folder. This is a tricky
business, as a special bell has been implemented, that the project
folders have to deal with the situation that the contents of several
conceptual folders are in the same digital folder on the hard disk. A
complicated filtering process is necessary to straighten that out. To be
able to use the properly functioning modules in the core, there must be
a text file in the JAR files that indicates the mapping from the
interface to the implementation. This file was missing in two modules
and has been added. It was found that the modules can not be loaded if
the library JARs they use are not referenced in the core. Therefore, the
necessary JARs have been added to the pom.xml of the core module.