Move the mapping of the JHove result values to the global configuration
How it works:
The bivalent mapping can be determined in the configuration file. An
example was added.

Implementation:
The configuration parameter is generated from the two-valued result and
looked up in the configuration file. From this the result is formed,
which is returned to production. The original state is added to the
messages.