Enables you to enable validation for folders
How it works:
To turn on validation for the contents of a folder, switches appear on
the task property page. By clicking on a switch, the function can be
switched on or off.

Implementation:
A new table validationFolders (task × folder) is created. This saves the
variable validationFolders in the task. The taskEditDetails.xhtml page
has been enhanced with switches to enable and disable the validation in
a step. The functionality for this has been implemented in the class
TemplateForm. The same functionality for generating images has also been
moved here, and duplicate code has been summarized. The GeneratorSwitch
class has been renamed to FolderProcessingSwitch because the switches
are used equally for both types of processing. The tests were adapted or
supplemented.