Implements a lock system for file access
How it works:
There are four different types of locks that are enforced. The complete
explanation can be found in the package-info.java.

Implementation:
The locks are implemented in the file management interface. There are
four varieties of it to deal with all sorts of cases. The management is
divided into three parts: the lock management, the stream management and
the immutable read file management. The lock management manages the
granted locks. The stream management manages the open streams through
the locks. The immutable read file management manages the temporary
files needed for the immutable read lock.