Automatically generate images in a task
Functioning:
When a task that has content folder generation enabled is completed, or
when an automatic task runs, the requested subfolders are generated.

Implementation:
The task script thread, which always runs when completing a task or
during automatic tasks, has been expanded to cause the images to be
generated as well. The corresponding method was installed in the
TaskService.