Add the domain to the meta-data class
As part of the refactoring process to extract the incorrectly
implemented service loaders for the METS module into data classes for
the API, the domain is taken from the
org.kitodo.dataformat.access.Metadata class to the
org.kitodo.api.Metadata class to use this as a data class for meta-data.

In Production, meta-data can be stored in five different areas that
describe different concept spaces. The different areas are represented
in the METS file as different XML elements.