Add data format API classes
As part of the refactoring process to extract the incorrectly
implemented service loaders for the METS module into data classes for
the API, the data classes that will be created during the extraction are
subsequently taken from the development and transferred to the API
beforehand, so that these can be used as data classes for the data
formats (METS, and possibly more in the future).

Please note that these classes are dead code as long as they have not
yet been extracted from the service loader.