Add a class for the linked structure
Next we need a class for the linked structure. It must have three
properties: the URI of the linked process, the title of the linked
structure, and the order. The order is a number that serves to place
other structures together with them in the correct order in the same
parental structure. After a link has been made, the order value no
longer affects the order.