Adds a function to resolve an URI in a link in a METS file
When loading a task, the METS access class must be able to resolve URIs
it finds. We extend the read() method with a reference to a function
that does this. This must also be done in the interface and in the
service loader. In the place where the service loader is called, this
method must be implemented. It is important to ensure a correct lock
management. To do this, the METS access class must specify in the
request to resolve a URI whether it may need to be written again later.
This may be the case for links down if they are removed by the user. For
links up this is never the case.