Add reading of child references
When building the child structure, it is checked whether the child
structure has a METS pointer to another METS file. If so, a linked
structure is created. In doing so, it is checked whether the back
reference actually leads back to the current METS file in order to avoid
damaging foreign processes in case of errors.