Show linked structures in the meta-data editor
In order to display superordinate structures above the root node of the
actual structure tree and that the whole looks the same, a list of trees
is generated. Each tree has only one node for a higher-level structure;
the last tree is the actual structure tree for the document.

The JSF code internally uses the legendary DocStructInterface.
Therefore, to show the links, the property whether it is a link, must be
included in the legacy interface and the implementing helper classes.
Since it lacks a separate object for the tree nodes, the helper class
MetadataProcessor must be extended by a method that calls the isLink()
function on the data object.