Re-implements the validation
The new validation has been re-implemented based on the concept of old
validation. Access to the document ID had to be added in order for it to
be validated. In several METS data objects toString() methods were added
in order to be able to display them in text form. The ruleset interface
had to be extended to access the frequencies values.