Improves the memorization of METS internal ID
Remembering the METS internal ID did not work everywhere. This is a bit
complicated because it has to be stored in the back of the data class
without any explicit storage space in the further. So we have to build
subclasses for storage and use. Also taken into account is the case that
the returned class from the data model is not a subclass.