Make the locking in the correct manner
Previously, locking METS files was only sloppy while editing: Before
reading or before writing, an exclusive lock was requested and then
discarded. The fact that two users do not get each other in the way was
ensured by the logic of the front-end. In the context of hierarchical
processes, the locks are of crucial importance. Therefore, locking must
now be done carefully. In particular, when opening a process all
necessary locks must be obtained and kept until closing.