Fix randomly failing selenium test
I ran the selenium tests locally and discovered that the try-catch-block in these methods was always called. Without them the test `addUserTest()` would fail completely.

My guess is that the webdriver tries to act on these elements although they are not ready yet by a tiny fraction of a second.

Therefore, `addUserTest()` “randomly” fails. Subsequent tests will fail as well because the browser throws an alert to warn you about unsaved input when leaving the page.

This change fix the test for me reliably.*

*Now something different will break.