Let the ruleset module know about metadata
This makes a lot of things easier and more understandable in the code,
and is a prerequisite for further development.

In the past, it was first decided to make the individual modules
independent of another, which also means that the ruleset module was
independent of the metadata that it processed. In the meantime, that has
become more relaxed, metadata is part of the API and can be used in
modules. So, this now introduces metadata into rule set modules, makes
it easier to use and more understandable, and gets rid of complicated
type parameters.

In the future, it can also be resolved that complex metadata could not
be mapped if not defined in the ruleset, because the ruleset module
could not be aware of it. Now it is, so that this problem, too, is easy
to solve in a continuation of this development.