Update KitodoServiceLoader to work with Java v11
Replaces the old class loading "hack" with a new similar class loading "hack", which also works with Java v11. Instead of adding jars to the webapp class loader, new independent class loaders are created and chained together. Similar to before, this is not a safe way of dynamically loading jar files. In the future, a dedicated plugin framework should be used instead, e.g., OSGi with Apache Felix.