Catch exceptions on corrupt data
In the case of corrupt data, e.g. damaged METS files, RuntimeException can be thrown out of the APIs used. If these are not logged, you will look in vain for the problem, the indexing just “gets stuck”. In this case, as an exception of the general rule, RuntimeException should be intercepted so that it is visible in the log file, so that the cause can be debugged. This is especially important when re-indexing, if otherwise nothing will go on.