Improve performance by not storing the full list of processes for each project in elastic search.
This list was only indexed in order to determine whether a project has any processes attached, such that the delete project icon could be enabled or disabled. This improves the indexing performance dramatically for large projects, because rendering and parsing long lists of thounsands of processes as json requires a lot of resources.

Requires re-indexing to take effect, since elastic mapping was changed.