Make all columns of the process table sortable.
Columns that were previously calculated on-the-fly are now added to the elastic search index and calculated at indexing time instead, e.g., the user name of the user that last worked on a task of the process.

The code that calculated these information has been moved to the Kitodo-DataManagment project in order to be available at indexing time.