Changes to search (#6587)
* Changes to the project search

The project search is now only performed in the database. The search
index field "searchProject" has been deleted. Searching with the
"project:" search field is now also possible using the SQL "LIKE"
operator. If a "*" is entered as the symbol, a left-truncated search is
also possible, for example, "project:*eitungen." The "project_loose:"
field always searches with a right-truncated search.

* Changes to task search

The task search is now performed only on the database. The search index
field "searchTask" has been deleted.

* Add missing file

* Fix checkstyle

* Fix checkstyle